le_role: Prevent the immediate reconnection in pair_adapter_to_device
Per the lab result, ~99% Floss is able to reconnect to the peer right
after disconnection because the peer is still advertising. And this
causes false positive - The tests only pass in this 99% cases; When
pair_adapter_to_device worked as expected, they actually fail and this
is the reason of the 1% flakiness on the dashboard.

This patch disables the peer advertising before disconnection and fixes
some test bugs revealed by this change:
- nearby_sender_role_test needs to be revised because per Floss
  configuration we wouldn't see the adv report in the btsnoop log.
- In receiver and secondary role tests we need to call
  ConnectAllEnabledProfiles right after test_connection_by_device
  because Floss passively doesn't connect to the HOG profiles for
  incoming connection.

BUG=b:325762700
TEST=bluetooth_AdapterLEHealth{.all_floss,} and checked btsnoop
TEST=bluetooth_AdapterLLTHealth.llt_health

Change-Id: Ia5ec0aed07804a46abad6cf3da618297fb06faaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5300085
Tested-by: Hsin-chen Chuang <chharry@google.com>
Auto-Submit: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>