FAFT: Ignore Google Servo as input devices
The class firmwareCheckKeys is currently taking the last keyboard input
device to monitor for key press input events. However, the order of
input devices is essentially random, so there's no guarantee that the
final keyboard input device is what the user is inputing key press
events to. Specifically, Google Servo devices are presented as keyboard
input devices, so if that is the final input device that's enumerated,
the test will fail due to not capturing any key press events.

To fix this, skip keyboard input devices with the string "Google Servo"
in the name.

BUG=b:322852586
TEST=test_that --autotest_dir ~/chromiumos/src/third_party/autotest/files \
    -b guybrush-kernelnext --args="servo_port=<port>" <IP> firmware_FAFTSetup

Change-Id: I3d4bfd2a7903745ee33ffab00967d8146d3c14e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5313387
Tested-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Commit-Queue: Tim Van Patten <timvp@google.com>