Rename USB3 control file.
It looks like some/all of the control file needs to match the exact name
of the tast group for the tests.

Since I'm not sure what needs to be updated, just update all the
relevant fields (filename, NAME attribute) to exactly match the TAST
group which is used for the tests (despite the continuous execution not
being related to bringup).

BUG=b:327251736
TEST=Tested on local DUT using:
     test_that --autotest_dir=<DIR> localhost:2022 \
     tast.typec_usb_bringup
     and validated that the tests run with the specified
     arguments.

Change-Id: I780c6a8182001dc4f52b1d742a0a695672454e14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5336462
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>