storage: check qual run ID last
in executions where the qual run ID was not populated, the prior
ordering silently did not populate the part_id field. We want to
populate two separate try catches such that the qual_run_id is not
dependent on the presence of the part_id and vice versa.

BUG=b:317989015
TEST=test_that, ensure that the run id is in keyval

Change-Id: I4d43d300c5e4844ad4d5a7e74bef7597ad60258a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5340891
Auto-Submit: Parker Holloway <jparkerh@google.com>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Parker Holloway <jparkerh@google.com>
Commit-Queue: Alexis Savery <asavery@chromium.org>