cr50_test: improve update from efi to dbg image
The test rolls back to the efi image. GSC rollback is triggered by a
reboot counter in long life scratch. Triggering a power-on GSC reset
will clear the counter and switch back to the DBG image. Use that on
boards with the gsc_reset signal to remove one update from the process.

Retry ccd updates while holding the device in reset. It's possible the
device is in a reboot loop and that's interfering with updates. Try ccd
updates with warm reset and cold reset asserted.

BUG=b:329287560
TEST=run test on strongbad

Change-Id: If207bdd3c64ff335aab047199843cd0cc8417840
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5385892
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>