cr50_test: use clear_rollback switch back to the DBG image
If the image has the clear_rollback command, try to use it to switch
back to the dbg image. It'll clear the rollback counter and reset. This
can help reliably switch from the EFI image back to the DBG image. It
doesn't use anything other than the cr50 console, so it should work on
all boards.

This command is only in cr50 EFI images. Ti50 doesn't need it since the
gsc reset signal is connected to the servo header on all ti50 devices.

BUG=b:329287560
TEST=none

Change-Id: I188ddb010b524416f0afc50783387582e9cfefe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5393250
Auto-Submit: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Aadit Maheshkumar Modi <modia@chromium.org>
Commit-Queue: Aadit Maheshkumar Modi <modia@chromium.org>