tast: Prevent Tast 2nd class runs of bluetooth*
Tast and Tauto bluetooth tests share suite names and both will
continue to be supported. Tast bluetooth tests are only to be
run via TFC moving forward. This deletes the Tast 2nd class
control files to prevent them from running when we are just tying
to run the Tauto tests.

Note: The control.bluetooth_cross_device control file is kept
for now as we still need to update the test to be able to support
TFC with the companionduts arg.

BUG=b:331255988
TEST=None

Change-Id: Iec0db92cd7391f6e86251f562d119ba1fd20a800
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5422029
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Terry Yan <yichiyan@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>