power: replace FDL short link with full link.
Tests became flaky due to failing to download parallax. This may be a
result of FDL links are being deprecated soon. Replacing the short link
with the full link for now to see if this resolves the problem.

BUG=b:333480576
TEST=test_that --board=<BOARD> --fast --debug <DUT> power_Idle

Change-Id: I3d7f055af09ef41088a10b06a2ca03a69eb41f73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5445304
Auto-Submit: Jingyuan Tu <zactu@google.com>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Jingyuan Tu <zactu@google.com>