fingerprint: Fix section extraction using 'dump_fmap'
Previously, '/dev/stderr' was passed to 'dump_fmap' and it was
interpreted as a section. It was obviously wrong, but it worked because
'dump_fmap' silently ignored missing sections.

CL:5408907 fixed that behavior. Now 'dump_fmap' fails if section was not
found, which showed the problem.

BUG=b:333396118
TEST=Run fingerprint test suite

Change-Id: I32c113412523b600cac98db39e6ed7aeaeab8e96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5458434
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>