telemetry_AFDOGenerate: upload to staging buckets on staging builds
It's unclear to me how to funnel information about the staging-ness of a
builder directly down to this test. In general, the current structure of
this test (read: having it upload artifacts directly to gs://) is
probably not a recommended practice.

The simplest way to work around this is likely to autodetect
staging-ness. This can be done from the name of the bot that generated
the image.

BUG=b:335653304
TEST=None

Change-Id: Ic2d7d627809ba15ef5b4554274d3e7e76551f246
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5466183
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Saswat Padhi <spadhi@google.com>
Auto-Submit: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>