Remove 'dump_vpd_log' usage
For several milestones now, the `vpd` tool has leveraged libvpd, such
that it is fast, and consistent with all other VPD writers. The stateful
filesystem cache is now obsolete, and we're preparing to remove it.

NB: I don't touch login_OobeLocalization here, because it's more complex
(poking at the inner details of the cache). And anyway, it's disabled at
the moment.

BUG=b:77594752
TEST=CQ

Change-Id: I25c52a68efb1dd45a8f21e1cb7128c5bc1d1f91c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5474315
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>