Bluetooth: Enable Chrome for SA Alias Tests
This change modifies the initialization of bluetooth_sa_health to
conditionally start Chrome when running the alias test or the entire
test batch. This is necessary because BlueZ's alias is set by Chrome,
and the Floss feature flag is required because some boards are default
to run Floss.

BUG=b:345673899
TEST=rm /var/lib/bluetooth/{MAC_ADDR}/settings then run
sa_valid_alias_test[.floss] and verify it passes.

Change-Id: Ice057cd71e856eb192fc023943df459c75001d86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5619796
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Tested-by: Yun-hao Chung <howardchung@chromium.org>
Commit-Queue: Yun-hao Chung <howardchung@chromium.org>