network_WiFi_BluetoothScanPerf: Migrate from BluetoothDevice to Facade
The test starts failing because we've switched some devices' default BT
stack to Floss. Migrate to Facade so we're able to switch the stack if
the desired one is not running. This patch now sets the desired stack to
Floss.

BUG=b:349712244
TEST=network_WiFi_BluetoothScanPerf.11a on brya (Floss default)
                                       and zork (BlueZ default)

Change-Id: Id5cfc0ab0b0ce15899d196c23def351fe37bd4b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5736175
Auto-Submit: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Tested-by: Hsin-chen Chuang <chharry@google.com>