floss: LE HID: Wait for HID device created
Waiting for 1s is sometimes not enough for the HID device to be created
after bonded, and it has been verified that the bottleneck is on the
btpeer side. This patch changes the 1s-wait to a poll to improve the
success rate.

BUG=b:356052972
TEST=bluetooth_AdapterLEHealth.le_{keyboard,mouse}_reports.floss

Change-Id: I17ed5e761312f23da9844c5b68c18610e9a083ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5743145
Tested-by: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Auto-Submit: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>