chromium: Don't hard-code 'out/Release/' as the build-dir
As part of crbug.com/355218109, we're changing the build-dir names of
builders across chrome's infra. This has smoked out instances where
tests have incorrectly assumed "out/Release" is the name of the build
dir.

This fixes one such assumption in the chrome->skylab pipeline by
taking the first two dirs of the "exe_rel_path". This isn't perfect, as
the build-dir might not always be nested two dirs in the checkout.

But this is far safer than assuming "out/Release". Properly propagating
the build-dir through to skylab would likely require a mysterious
amount of plumbing. So the proper-fix is left as a follow-up, as this
change as-is is sufficient to unblock the work in crbug.com/355218109.

BUG=chromium:355218109
TEST=whatever runs in presubmit on this repo

Change-Id: I87a72e1dcfa3addddcd8634ec41b5bd333005e42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5793917
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>