telemetry_AFDOGenerate: refactor ARM USR2 loop; have perf show life
Existing code used perf's PID to determine if perf was ready to sample.
This behavior is suspected to be causing flakiness in the event that
perf takes a bit to start up: b/369803532#comment5

Swap to a "handshake" to remove this race, and a USR2 loop that will
create more smoke if it happens to die unexpectedly.

BUG=b:369803532
TEST=staging-afdo-generate-orch

Change-Id: I4a93a4189d26227a185b15168cd084f8dcd7772d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5898313
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>