bluetooth: Always open evtest in the background
On autotest environment, UHID only opens for a brief second before it
is closed again. That is because stopping UI causes UHID_CLOSE to be
sent for some reason.

Since aosp/3235675, we disconnect the BT device if it takes too long
to open UHID. This patch doesn't bode well with the behavior of UHID
in autotest, causing floss to disconnect the peer devices.

To simulate the behavior of the normal use cases where UHID doesn't
send UHID_CLOSE, this CL opens an evtest, therefore acts as a listener
of the input event, thus preventing UHID_CLOSE from being sent.

BUG=b:371680523
TEST=autotest

Change-Id: Ic286b04494615c5a88ead596e334eba0ef225c9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5912318
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>
Tested-by: Archie Pusaka <apusaka@chromium.org>