chrome_ti50: fail test if "externally driven!" message shows up
"externally driven!" means that the external WP signal doesn't match the
WP state ti50 is trying to set. It shouldn't happen. Servod sets the
ti50 WP state before driving the signal.

Fail the test if "externally driven!" shows up in the logs, so we can
see what tests trigger the issue.

This CL may need to be reverted if it causes a lot of tests to fail.

BUG=none
TEST=none

Change-Id: Id5c971ab998f423628d879e0bbe69139446ba5e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5916831
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Jett Rink <jettrink@google.com>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>