bluetooth: refactor chameleon restart process
Refactor the test to avoid a race condition in the restart process.

The original test process has a flaw:

It restarts the chameleond.
It immediately sends an RPC request to verify the server restart.
This sequence can cause a race condition. If the RPC request arrives
before the server fully restarts, it may receive a response from
the old chameleond instance, leading to a false positive.

This CL contains the test changes. The corresponding Chameleond changes
are in CL 5937057.

BUG=b:369543317
TEST=sr_peer_wake_le_hid.floss

Change-Id: Id08ab2ae4df07e61f2960d631264bfc0de5d5fa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/5936984
Reviewed-by: Zhengping Jiang <jiangzp@google.com>
Commit-Queue: John Lai <johnlai@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: John Lai <johnlai@google.com>