autotest-server-tests: Remove ping check after chameleon reboot.
The polling is depending on the port forwarding and SSH connection re-establishment by ChameleonConnection class, which contains quite complicated retry mechanisms.

Considering the purpose of this test is simply rebooting the chameleon, simply wait for a while and ignore this part. If the chameleon is not awake, it shall be detected in the follow-up test.

Added 10s sleep time before leaving the test.

BUG=b:377899099
TEST=local tested

Change-Id: I07878a60d42a09b3761c7639b607ef8b650240a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6003755
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Tested-by: YJ Lee <yunjunlee@chromium.org>
Commit-Queue: Kalin Stoyanov <kalin@chromium.org>