bluetooth: test host privacy setting is restored after reboot
When LL privacy is manually enabled, the host address policy is changed
to RPA. If a device is paired after that, the device only recognizes the
host with the same address policy. This test is to make sure the address
policy is kept the same after reboot, and the device paired before can
still reconnect. Even though the LL privacy is disabled after reboot.

BUG=b:384597542
TEST=run LL privacy suite

Change-Id: Ib93745bc1ef89c931e3b53d63a0c8a38e2a349b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6177786
Commit-Queue: Zhengping Jiang <jiangzp@google.com>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Tested-by: Zhengping Jiang <jiangzp@google.com>