Initialize target_build and target_cr to avoid UnboundLocalError
The variables target_build and target_cr are not initialized before
the loop that iterates through the CSV rows. If the board name is not
found on any row, these variables remain undefined, resulting in
an UnboundLocalError when checking for their value in the following if
condition.

Initialize target_build and target_cr to None before the loop to ensure
that they always have a value, even if no matching row is found.

BUG=b:391964986
TEST=autoupdate_Basic.control.m2n.full

Change-Id: I6a09192844ffcee8cddd4545f4b636f14b5e9adf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6227799
Tested-by: Joaquin Marquez <joaquinmarquez@google.com>
Commit-Queue: Joaquin Marquez <joaquinmarquez@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>