Improve logic to get serving stable build
Currently, the logic to get the latest (or oldest) serving stable build
is not working as intended when there is no data for stable builds in
the serving builds page/CSV and in paygen.json. This is happening when
the board is in extended support. Additionally, the current logic is
comparing version numbers lexicographically, which is causing some very
old builds to be picked up as the “latest” build, since for example
9202.64.0 is greater than 16093.104.0, when comparing lexicographically.

Considering LTS and LTC builds, in addition to stable, should allow the
tests to find the latest stable build to flash the device with. Also,
update logic to use packaging.version utils to properly parse and compare
version numbers.

BUG=b:412911937
TEST=Ran autotest_Basic.m2n.full tests for all affected boards

Change-Id: I336f6e29c512b8aa23de7de3eda30271e2089e67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6506761
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Joaquin Marquez <joaquinmarquez@google.com>
Tested-by: Joaquin Marquez <joaquinmarquez@google.com>