autotest: add sync at the end of platform_CryptohomeFio
The test flakes because of unmount timeout inside cryptohome cleanup.
Based on the log analysis, the hypothesis is that namespace mounter
takes too long to unmount because it needs to sync the data inside the
volume. Syncing the FS within the test doesn't have a timeout, which is
pretty agressive inside cryptohomed.

BUG=b:418425117
TEST=test_that <DUT> platform_CryptohomeFio

Change-Id: I2d5e91c27a1af2acd3ccdca01f7af382934ebf0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6636329
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>