floss: Support unstable Aflags selection
Add the floss_use_unstable_aflags arg in BT test batches setup. When
it's set to True by the control file, the unstable Aflags would be
applied.

The unstable Aflags are the flags that are shipped to the Android users
by the quarterly release; They will likely be removed in the next
release and we enable them in the ChromeOS Beta/Dev channel for testing/
stabilizing. See https://crrev.com/c/6817812 which adds the list.

For automated tests we will test both the "unstable Aflags applied" and
the "not applied". The existing tests will be the "not applied"; While
the control files for the "unstable Aflags applied" will be added soon.

BUG=b:435354365
TEST=bluetooth_AdapterSAHealth.sa_valid_address_test.floss.unstable_aflags
TEST=bluetooth_AdapterSAHealth.all_floss.unstable_aflags
TEST=bluetooth_AdapterQuickHealth.AVL.all_floss.unstable_aflags

Cq-Depend: chromium:6817812
Change-Id: I20ac96baf7f135ce162e6176c2c6574bcc6c4719
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6819945
Reviewed-by: John Lai <johnlai@google.com>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Tested-by: Hsin-chen Chuang <chharry@google.com>