EP: Extend the sleep in wait_until_device_ready
The sleep should at least allow the service discovery to be finished.
The failure rate was not so high, so only extending a little bit.

Normally we should avoid this kind of sleep because it's unreliable and
tends to cause flakies, but for the expected_fail case it's hard to
determine "ready" as it will never become ready.

BUG=b:442760412
TEST=bluetooth_AdapterEPHealth.ep_reconnection_ble_hid_service_not_in_allowlist.floss.unstable_aflags

Change-Id: Idb9cc35033e271b66e94c5f8a42cef6dd13f096b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/6937628
Auto-Submit: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Yun-hao Chung <howardchung@chromium.org>
Tested-by: Hsin-chen Chuang <chharry@google.com>